Extracts Assembler into a separate module
