+++ BoneCentral/Peripherals/Peripherals.Assembler/include/ScriptLogger.h
