+++ BoneCentral/Peripherals/Peripherals.Assembler/src/ScriptLogger.cpp
