+++ BoneCentral/Peripherals/Peripherals.SerialAdaptors/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.9)
-project(Peripherals)
+project(Peripherals.SerialAdaptors)
 
 set(SerialAdaptor_Includes "${CMAKE_CURRENT_SOURCE_DIR}/include")
 include_directories(${Serial_Adaptor_Includes})
