+++ BoneCentral/Peripherals/main.cpp
@@ -1,0 +1,13 @@
+//
+// Created by Nathan Copier on 4/20/2016.
+//
+
+#include <Assembler.h>
+
+int main(int argCount, char** arguments) {
+
+    App_Start(argCount, arguments);
+
+    return 0;
+}
+
